rename & re-arrange some docs components, add autoscroll to docs, buttonBookmarks, docs styling, navigation flow type
