+++ app/src/docs/components/ButtonBookmark/buttonBookmark.pcss
@@ -1,0 +1,42 @@
+@import 'variables.pcss';
+
+.root {
+  flex-basis: calc(50% - 8px);
+  min-height: 176px;
+  background: #FFFFFF;
+  border: 1px solid #D8D8D8;
+  border-radius: 4px;
+  padding: 1.5em;
+
+  h5 {
+    font-size: 20px;
+    font-family: 'IBM Plex Sans', sans-serif;
+    font-weight: var(--semiBold);
+    letter-spacing: 0;
+    line-height: 28px;
+  }
+
+  p {
+    color: #525252;
+    font-size: 14px;
+    letter-spacing: 0;
+    line-height: 20px;
+  }
+
+  a {
+    font-weight: var(--medium);
+    color: var(--blue);
+    text-decoration: none;
+
+    &:hover {
+      color: var(--blue);
+    }
+  }
+}
+
+@media (--md-down) {
+  .root {
+    flex-basis: auto;
+    margin: 1em 0;
+  }
+}
