+++ app/src/docs/components/DocsLayout/Navigation/navigation.pcss
@@ -52,3 +52,14 @@
     }
   }
 }
+
+@media (--md-down) {
+  .navList,
+  .subNavList {
+    background-color: #F8F8F8;
+  }
+
+  .subNavList {
+    display: none;
+  }
+}
