+++ app/src/docs/components/DocsLayout/docs.pcss
@@ -1,5 +1,6 @@
 .layout {
   display: flex;
   flex: 1;
+  padding: 2.5em 0;
   color: #323232;
 }
