+++ app/src/docs/components/GettingStartedPage/index.jsx
