+++ app/src/docs/components/IntroductionPage/introduction.pcss
@@ -1,0 +1,49 @@
+@import 'variables.pcss';
+
+.streamrBanner {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 280px;
+  width: 100%;
+  margin: 2em 0;
+  background-color: var(--blue);
+  border-radius: 2px;
+
+  .logo {
+    width: 5em;
+    height: 5em;
+  }
+}
+
+.forwardArrow {
+  position: absolute;
+  top: 5px;
+  right: 0;
+  transform: rotate(180deg);
+  width: 12px;
+  height: 12px;
+
+  path {
+    stroke: var(--blue);
+  }
+}
+
+.buttonBookmarks {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  margin-bottom: 2em;
+}
+
+.link {
+  position: relative;
+  padding-right: 20px;
+
+  &:hover {
+    .forwardArrow {
+      transform: translate(4px, 0) rotate(180deg);
+      transition: all 0.3s ease-out;
+    }
+  }
+}
