+++ app/src/docs/flowtype/navigation-types.js
@@ -1,0 +1,5 @@
+// @flow
+
+export type NavigationLink = {
+    [string]: string
+}
